+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.1.8: 2022-08-18
 
 * Add support for renaming ready-made blocks
+* `newblock`: Add hero-big-image
 
 ### 1.1.7: 2022-08-18